third_party/blink/web_tests/platform/generic/fast/events/anchor-empty-focus-expected.txt

This tests that focus() and sequential focus navigation work on empty anchors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement is emptyAnchor1
PASS document.activeElement is emptyAnchor2
PASS successfullyParsed is true

TEST COMPLETE
